Modificado: Tablas "unit_types", "shopping_list" y "products_has_shopping_list" - Se cambia el tipo de la columna "id" por el tipo "BIGINT". - Se actualiza las entidades de java y todas las clases donde se utiliza el campo ID. - Agregado script para flyway.

+++ src/main/java/dev/nmarulo/depensaapp/app/products/dtos/SaveShoppingListProductReq.java
@@ -7,10 +7,10 @@ public class SaveShoppingListProductReq {
     
     private Long productId;
     
-    private Integer shoppingListId;
+    private Long shoppingListId;
     
     private Integer unitsPerProduct;
     
-    private Integer unitTypeId;
+    private Long unitTypeId;
     
 }
