+++ src/main/java/dev/nmarulo/depensaapp/app/unitytypes/UnitTypeRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UnitTypeRepository extends JpaRepository<UnitType, Integer> {}
+public interface UnitTypeRepository extends JpaRepository<UnitType, Long> {}
